Add  Sorting Algo implementation in Java
Add a Java implementation of the Bubble, Selection and Insertion Sort algorithm. 
Bubble Sort - works by repeatedly stepping through the list, comparing adjacent elements, and swapping them if they are in the wrong order. This process continues until no more swaps are needed, meaning the list is sorted. The algorithm has a time complexity of O(n²), making it inefficient for large datasets, but it is simple and useful for small or nearly sorted arrays.
Selection - The algorithm iterates through the list, repeatedly finding the minimum element from the unsorted part and moving it to the sorted part.
Insertion - This algorithm iteratively builds a sorted array by picking one element from the unsorted part and inserting it into its correct position within the sorted part.